+++ waterfall-cad-examples/app/Main.hs
@@ -6,6 +6,7 @@ import GearExample (gearExample)
 import FilletExample (filletExample)
 import RevolutionExample (revolutionExample)
 import SweepExample (sweepExample)
+import OffsetExample (offsetExample)
 import TextExample (textExample)
 import Waterfall.IO (writeSTL, writeSTEP)
 import qualified Waterfall.Solids as Solids
@@ -26,6 +27,7 @@ exampleOption = OA.flag' csgExample (OA.long "csg" <> OA.help "example from the 
                 OA.flag' filletExample (OA.long "fillet" <> OA.help "demonstrates adding fillets to an object" ) <|>
                 OA.flag' revolutionExample (OA.long "revolution" <> OA.help "demonstrates revolving a path into a solid" ) <|>
                 OA.flag' sweepExample (OA.long "sweep" <> OA.help "demonstrates sweeping a shape along a path" ) <|>
+                OA.flag' offsetExample (OA.long "offset" <> OA.help "demonstrates offsetting the surface of a shape" ) <|>
                 (OA.flag' gearExample (OA.long "gear" <> OA.help "generate an involute gear") <*>
                  (OA.option OA.auto (OA.long "thickness" <> OA.help "gear depth") <|> pure 1.0) <*>
                  (OA.option OA.auto (OA.long "module" <> OA.help "gear module") <|> pure 5.0) <*>