+++ waterfall-cad/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to the
 
 ## Unreleased
 
+### Fixed
+
+- Handle offsetting by zero (return the unmodified shape)
+
 ## 0.1.1.1 - 2023-12-12 
 
 ## 0.1.1.0 - 2023-12-12